requirements.txt
# [Windows] - PowerShell
# Get-Content requirements.txt | ForEach-Object {
#     if ($_ -match "\S" -and -not $_.StartsWith("#")) {
#         Write-Host "Installing $_..." -ForegroundColor Cyan
#         pip install $_
#     }
# }

# [Linux]
# while read line; do
#   # Skip empty lines and comments
#   if [[ -n "$line" && ! "$line" =~ ^[[:space:]]*# ]]; then
#     echo "Installing $line..."
#     pip install $line
#   fi
# done < requirements.txt

numpy==1.26.4

albumentations>=2.0.5
clearml>=1.18.0
nltk>=3.9.1
lightning>=2.4.0
lightning-flash>=0.8.2
kaggle>=1.6.17
pandas>=2.2.3
pydicom>=3.0.1
omegaconf>=2.3.0
torchmetrics>=1.6.1
transformers>=4.49.0
tensorboard>=2.18.0
wandb>=0.19.2
rich>=13.9.4
opencv-python>=4.10.0
seaborn>=0.13.2
scikit-learn>=1.6.0
scikit-image>=0.25.0
segmentation-models-pytorch>=0.4.0
mplcyberpunk>=0.7.6
